moved functions from compas_extra.py to package level
