+++ src/compas_timber/ghpython/components/CT_ShowTopologyTypes/code.py
@@ -4,7 +4,7 @@ from compas_rhino.conversions import point_to_rhino
 from ghpythonlib.componentbase import executingcomponent as component
 
 from compas_timber.connections import JointTopology
-from compas_timber.utils.compas_extra import intersection_line_line_3D
+from compas_timber.utils import intersection_line_line_3D
 
 
 class ShowTopologyTypes(component):
